Fix possible integer overflow

+++ src/layer.cpp
@@ -245,7 +245,7 @@ bool Layer::save(QString filename)
 {
   // Check if canvas is largely transparent. If so, don't save it
   const QRgb *canvasBits = (QRgb *)canvas.constBits();
-  quint64 noOfPixels = canvas.width() * canvas.height();
+  quint64 noOfPixels = static_cast<long>(canvas.width()) * canvas.height();
   bool transparent = true;
   for(quint64 a = 0; a < noOfPixels; a = a + 10) {
     if(qAlpha(canvasBits[a]) != 0) {
